Moved all controllers to the controller folder
