+++ src/main/java/controller/TaskEventAutoController.java
@@ -1,4 +1,4 @@
-package main.java.use_case;
+package main.java.controller;
 
 import java.time.LocalDateTime;
