+++ src/main/java/controller/TaskToEventConsoleController.java
@@ -1,8 +1,5 @@
 package main.java.controller;
 
-import main.java.use_case.TaskEventAutoController;
-import main.java.use_case.TaskEventManualController;
-
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Scanner;
