+++ src/main/java/use_case/TaskToEventAuto.java
@@ -1,4 +1,5 @@
 package main.java.use_case;
+import main.java.controller.TaskEventAutoController;
 import main.java.entity.Calendar;
 import main.java.entity.Event;
 import main.java.entity.Task;
